Minor Tester tweaks + wrote PackBits compressor
The Tester now exits when an existing output file cannot be closed. If this happens, it is most likely being used by another process, and needs to be dealt with manually.

GoodPackBits implements PackBits, which is an RLE compression algorithm from old Macintoshes. It has provisions for literal byte sequences, so it doesn't double the length of most files!